Update settings with new "period" parameter
A new setting is added, "period", to tell the script to collect the last x days of records from the SmartPTT database. It supports fractions of a day, i.e. to collect the last 12 hours of data then "period = 0.5"
Also, added some more details about other parameters and their use.